# docs/design/Endpoint/openapi-admin.yaml
openapi: 3.1.0
info:
  title: Sparr Admin Endpoints
  version: 1.0.0
  description: |
    Admin endpoint specification. Corresponds to the Admin Endpoint definition in `docs/architecture/components.md`.

    ## Authentication

    Auth routes (sign-in / sign-out / OAuth callbacks) are provided by the Auth adapter at
    `app/api/auth/[...auth]` (framework-managed via `providers.auth.handlers` delegation).
    The concrete URL scheme is adapter-defined and is not specified in this document.

    Business endpoints assume an authenticated principal via `getPrincipal()` and do not own
    sign-in or sign-out flows. `role=admin` is enforced per-route via `requireAdmin()` —
    not by individual endpoint sign-in logic.

    CSRF protection relies on the SameSite cookie policy (Auth.js-independent);
    no X-CSRF-Token header is required.

    All /api/admin/** paths require role=admin enforced per-route via requireAdmin().
    Operations touch ACCOUNT / SCENARIO / PERSONA / RESULT / SETTING via the Store.
    Player endpoints (`/api/player/**`) are documented in `./openapi.yaml`.
servers:
  - url: http://localhost:3000
    description: Local dev
...
paths:
...
  /api/auth/password-change:
    post:
      tags: [Auth]
      summary: Change the authenticated user's credential
      description: |
        Delegates to `providers.auth.changePassword()`.

        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Accept CredentialChangeRequest (currentCredential / newCredential)
        3. Call providers.auth.changePassword() — verifies currentCredential (401 if mismatch),
           validates newCredential (min 8 characters), and persists it
        4. Return 200
      operationId: postAuthPasswordChange
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialChangeRequest"
      responses:
        "200":
          description: Credential changed successfully
        "400":
          description: Validation error (e.g. new credential too short)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated or current credential mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Change failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/users:
    get:
      tags: [Users]
      summary: List users
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Build filter from query parameters (role / q)
        4. Call Store.accounts.list(filter)
        5. Return AdminUserSummary[] with 200

      operationId: getAdminUsers
      security:
        - SessionCookie: []
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [player, admin]
          description: Filter by role
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Partial-match filter on email or username
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserSummary"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Store retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Users]
      summary: Create a user
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Accept AdminUserCreateRequest (email / credential / username? / role)
        4. Call providers.auth.register() to validate, check for duplicates, and persist the new account
        5. Return the created AdminUserSummary with 201

      operationId: postAdminUsers
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreateRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserSummary"
        "400":
          description: Validation error (missing required field, credential too short, invalid email format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email address already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: |
        Session cookie set by the Auth adapter on sign-in. Sent automatically by the browser
        for same-origin requests. CSRF protection relies on the SameSite cookie policy —
        no separate X-CSRF-Token header is required.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth / Account ────────────────────────────────────
...
    CredentialChangeRequest:
      type: object
      required: [currentCredential, newCredential]
      properties:
        currentCredential:
          type: string
          description: Current credential for identity verification
        newCredential:
          type: string
          minLength: 8
          description: New credential (min 8 characters)
...
    AdminUserSummary:
      type: object
      required: [id, email, role, username]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [player, admin]
...
    AdminUserCreateRequest:
      type: object
      required: [email, credential, role]
      properties:
        email:
          type: string
          format: email
        credential:
          type: string
          minLength: 8
          description: Initial credential (min 8 characters)
        username:
          type: string
          description: Username (defaults to the local part of the email if omitted)
        role:
          type: string
          enum: [player, admin]
...